requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.9.3
blinker==1.8.2
cachetools==5.3.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
cryptography==40.0.1
distro==1.9.0
Flask==3.0.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.1.1
frozenlist==1.3.3
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
idna==3.4
ijson==3.2.3
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
langdetect==1.0.9
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.0
pdfminer.six==20221105
pdfplumber==0.8.0
Pillow==9.5.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
PyPDF2==3.0.1
python-dotenv==0.19.0
requests==2.26.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.31
tqdm==4.65.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==1.26.15
Wand==0.6.11
Werkzeug==3.0.3
yarl==1.8.2
youtube-transcript-api==0.6.1
